hotfix: fix remove/upload avatar on signup (#384)

+++ frontend/src/components/UserProfile/UserInfo.js
@@ -37,7 +37,11 @@ const ProfilePage = (props) => {
   };
 
   const handleAvtRemove = () => {
-    if (props.user.avatar.origin) {
+    if (
+      props.user.avatar &&
+      avt !== defaultAvt &&
+      avt === props.user.avatar.origin
+    ) {
       deleteAvatar({
         onError: (error) => {
           toast.error(error.message);
